requirements.txt
GDAL @ file:///D:/Projects/Python/datingsite/GDAL-3.4.3-cp311-cp311-win_amd64.whl
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.2
django-debug-toolbar==4.1.0
django-debug-toolbar-force==0.2
django-environ==0.10.0
django-filter==23.2
django-nine==0.2.7
djangorestframework==3.14.0
Flask==2.3.2
geographiclib==2.0
geopy==2.3.0
geos==0.2.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2==2.9.6
pygeos==0.14
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6